Clean conclusion tests for TIAC
- Move to a dedicated file
- Use new trait when possible
